Criando testes, inserindo imagem do ícone do google abrindo popup
